revamped install scripts

+++ build/scripts/040-nginx.sh
@@ -1,10 +1,22 @@
 #!/bin/sh
 
-# copy nginx files to correct location
-cp ../website/jameswhale /etc/nginx/sites-available/jameswhale
+# create nginx file if it doesn't exist
+[ ! -f '/etc/nginx/sites-available/jameswhale' ] && cat > /etc/nginx/sites-available/jameswhale << EOF
+server {
+    listen 80;
+    listen [::]:80;
+
+    server_name james-whale.com www.james-whale.com;
+
+    location / {
+        include uwsgi_params;
+        uwsgi_pass unix:/var/www/website_redux/jameswhale.sock;
+    }
+}
+EOF
 
 # make site available
-ln -s /etc/nginx/sites-available/jameswhale /etc/nginx/sites-enabled
+[ ! -e '/etc/nginx/sites-enabled/jameswhale' ] && ln -s /etc/nginx/sites-available/jameswhale /etc/nginx/sites-enabled
 
 # check for errors
 nginx -t && systemctl restart nginx || return
